refactor: moves "options.json" to the root directory
