+++ docs/usage-cli.rst
@@ -83,4 +83,4 @@ CLI Options Reference
 
 Remember you can always list all available arguments by ``dredd --help``.
 
-.. cli-options:: ../lib/options.json
+.. cli-options:: ../options.json
