+++ lib/CLI.js
@@ -14,7 +14,7 @@ const logger = require('./logger');
 const { applyLoggingOptions } = require('./configuration');
 const { spawn } = require('./childProcess');
 
-const dreddOptions = require('./options');
+const dreddOptions = require('../options.json');
 const packageData = require('../package.json');
 
 class CLI {
